+++ src/components/exercises/SelectExercise/SelectExercise.jsx
@@ -3,7 +3,7 @@ import {
 	IconButton,
 	ProgressDots,
 } from "@/components";
-import { exerciseActionButtonVariants } from "@/components/exerciseActionButtonVariants";
+import { exerciseActionButtonVariants } from "@/components/exercises/shared/exerciseActionButtonVariants";
 import {
 	Select,
 	SelectContent,